refactor(MultipartAssets): large/small variants names
