+++ src/script/components/MessagesList/Message/ContentMessage/asset/MultipartAssets/ImageAssetCard/ImageAssetLarge/ImageAssetSmall.tsx
@@ -33,15 +33,15 @@ import {
   infoOverlayStyles,
   infoWrapperStyles,
   loaderIconStyles,
-} from './LargeImageAsset.styles';
+} from './ImageAssetSmall.styles';
 
-interface LargeImageAssetProps {
+interface ImageAssetLargeProps {
   src?: string;
   metadata: ICellAsset['image'];
   isError: boolean;
 }
 
-export const LargeImageAsset = ({src, metadata, isError}: LargeImageAssetProps) => {
+export const ImageAssetLarge = ({src, metadata, isError}: ImageAssetLargeProps) => {
   const [isLoaded, setIsLoaded] = useState(false);
 
   const aspectRatio = metadata?.width && metadata?.height ? metadata?.width / metadata?.height : undefined;
